Add missing call to xf86SetCrtcForModes() and add pass through option
This fixes the problems with the crashes and other funnies. It was almost
entirely down to a missing call to xf86SetCrtcForModes(). With that added
the clock data appears valid and the card is working.

I've also added support for the pass through switching as that is useful
for debug when your text consoles are on the pass through display